// src/types/calendar.types.ts
export type NewCalendarData = {
	calendarInfo: {
		dates: {
			bs: {
				year: {
					np: string | null;
					en: string | null;
				};
				month: {
					np: string | null;
					code: {
						np: string | null;
						en: string | null;
					};
					en: string | null;
				};
				day: {
					np: string | null;
					en: string | null;
				};
				full: {
					np: string | null;
					en: string | null;
				};
			};
			ad: {
				year: {
					np: string | null;
					en: string | null;
				};
				month: {
					np: string | null;
					code: {
						np: string | null;
						en: string | null;
					};
					en: string | null;
				};
				day: {
					np: string | null;
					en: string | null;
				};
				full: {
					np: string | null;
					en: string | null;
				};
			};
		};
		days: {
			dayOfWeek: {
				np: string | null;
				en: string | null;
			};
			codes: {
				np: string | null;
				en: string | null;
			};
		};
		nepaliEra: {
			sakSambat: {
				np: string | null;
				en: string | null;
			};
			nepalSambat: {
				year: {
					np: string | null;
					en: string | null;
				};
				month: {
					np: string | null;
					code: {
						np: string | null;
						en: string | null;
					};
					en: string | null;
				};
			};
		};
	};
	tithiDetails: TithiDetails;
	panchangaDetails: PanchangaDetails;
	hrituDetails: {
		id: string | null;
		title: {
			np: string | null;
			en: string | null;
		};
		codes: {
			np: string | null;
			en: string | null;
		};
	} | null;
	auspiciousMoments: {
		sahits: Array<{
			id: string | null;
			title: {
				np: string | null;
				en: string | null;
			};
			description: {
				np: string | null;
				en: string | null;
			};
			times: {
				start: {
					np: string | null;
					en: string | null;
				};
				end: {
					np: string | null;
					en: string | null;
				};
			};
		}>;
		muhurats: Array<{
			periodName: string | null;
			duration: string | null;
		}>;
	};
	eventDetails: EventDetail[];
};

export type PanchangaDetails = {
	pakshya: {
		np: string | null;
		en: string | null;
	};
	chandraRashi: {
		time: {
			np: string | null;
			en: string | null;
		};
		endTime: {
			np: string | null;
			en: string | null;
		};
	};
	suryaRashi: {
		np: string | null;
		en: string | null;
	};
	nakshatra: {
		np: string | null;
		endTime: {
			np: string | null;
			en: string | null;
		};
		ghaPa: {
			np: string | null;
			en: string | null;
		};
	};
	yog: {
		np: string | null;
		en: string | null;
		endTime: {
			np: string | null;
			en: string | null;
		};
	};
	karans: {
		first: {
			np: string | null;
			endTime: {
				np: string | null;
				en: string | null;
			};
		};
		second: {
			np: string | null;
			endTime: {
				np: string | null;
				en: string | null;
			};
		};
	};
	season: {
		name: {
			np: string | null;
			en: string | null;
		};
		code: {
			np: string | null;
			en: string | null;
		};
	};
	times: {
		sunrise: string | null;
		sunset: string | null;
		moonrise: string | null;
		moonset: string | null;
	};
} | null;

export type TithiDetails = {
	title: {
		np: string | null;
		en: string | null;
	};
	code: {
		np: string | null;
		en: string | null;
	};
	endTime: {
		np: string | null;
		en: string | null;
	};
	display: {
		np: string | null;
		en: string | null;
	};
} | null;

export type EventDetail = {
	id: string | null;
	category: {
		name: string | null;
		id: string | null;
	};
	displayType: string | null;
	title: {
		np: string | null;
		en: string | null;
	};
	isHoliday: boolean;
	location: string | null;
	timing: {
		from: string | null;
		to: string | null;
	};
	details: {
		en: string | null;
		np: string | null;
	};
};